+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Releases are added to the
 ## [3.2.0] -- 2024-07-28
 
 - Remove support for Django 3.x (LTS is EOL)
-- Automatically add `crossorigin` attributes to tags with `integrity` attributes when necessary (and enabled)
+- Automatically add `crossorigin` attributes to tags with `integrity` attributes when necessary
 - Use `request.csp_nonce` from [django-csp](https://github.com/mozilla/django-csp) if available and configured
 
 ## [3.1.0] -- 2024-04-04